added words.py file
